+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Changes in October 2021
 
+-   Changed the parameter for takeScreenshot in RecentGraphScreenshotUtilities 
+    from filename to filepath.
+
 -   Added `isRequired` in `PluginParameter` with a getter and a setter, which 
     can be used to configure the required plugin parameters to mark as `*required`
 	in the swagger.